requirements.txt
python-telegram-bot[ext]>=21.6
pyrogram>=2.0.106
motor>=3.5
tgcrypto>=1.2.5
apscheduler>=3.10.4
